[ZODB4] setup: tests_require += zope.testrunner
Zope.testrunner is used to run ZODB tests (see e.g. tox.ini). It was
specified as explicit dependency in tox.ini->deps and in buildout.cfg.
However if in SlapOS we install just ZODB[test] zope.testrunner won't be
installed.

-> Fix is by moving zope.testrunner dependency into setup, so that
ZODB[test] installs it. This is the same as what ZODB5 does:

https://github.com/zopefoundation/ZODB/blob/5.6.0-14-g0eae10cd0/setup.py#L48