[ZODB3] Provide stub ZODB/persistent/BTrees/ZEO eggs for forward-compatibility with ZODB4 and ZODB5
This depend-only eggs have to be installed manually for
forward-compatibility with ZODB4 and ZODB5, so that e.g. installing
any egg that depends on 'ZODB' (not 'ZODB3') will not pull in ZODB5.
Example of those eggs are zodbtools, zodburi, ...

Unfortunately there is no way for this to work automatically - for
example there is 'provides' and 'provides-dist' setup keywords and
corresponding metadata, but no Python package manager actually supports
that:

    https://packaging.python.org/specifications/core-metadata/#rarely-used-fields
    https://www.python.org/dev/peps/pep-0314/#provides-multiple-use
    https://www.python.org/dev/peps/pep-0345/#provides-dist-multiple

    https://github.com/pypa/packaging-problems/issues/154   (obsoletes/provides not supported)